ver3
memperbarui direktori file.
bikin rapih.
