+++ app/controllers/OrderController.php
@@ -17,27 +17,53 @@ class OrderController extends Controller {
 
     public function checkout() {
         if ($_SERVER['REQUEST_METHOD'] == 'POST') {
-            $cart = [];
-            foreach ($_POST['jumlah'] as $id_menu => $jumlah) {
-                if ($jumlah > 0) {
-                    $cart[$id_menu] = [
-                        'jumlah' => $jumlah,
-                        'harga' => $_POST['harga'][$id_menu]
-                    ];
+            // Start output buffering
+            ob_start();
+            
+            try {
+                $cart = [];
+                foreach ($_POST['jumlah'] as $id_menu => $jumlah) {
+                    if ($jumlah > 0) {
+                        $cart[$id_menu] = [
+                            'jumlah' => $jumlah,
+                            'harga' => $_POST['harga'][$id_menu]
+                        ];
+                    }
                 }
-            }
 
-            if (empty($cart)) {
-                die("Keranjang kosong!");
-            }
+                if (empty($cart)) {
+                    throw new Exception('Keranjang masih kosong! Silakan pilih menu terlebih dahulu.');
+                }
 
-            $orderModel = $this->model('Order_model');
-            $result = $orderModel->processCheckout($_SESSION['user_id'], $cart);
+                $orderModel = $this->model('Order_model');
+                $result = $orderModel->processCheckout($_SESSION['user_id'], $cart);
 
-            if ($result['status']) {
-                $this->redirect(BASEURL . '/order/antrian/' . $result['nomor_antrian']);
-            } else {
-                die("Checkout Gagal: " . $result['message']);
+                if ($result['status']) {
+                    // Clear output buffer
+                    ob_end_clean();
+                    
+                    // Send JSON response
+                    header('Content-Type: application/json');
+                    echo json_encode([
+                        'status' => 'success',
+                        'message' => 'Pesanan berhasil dibuat',
+                        'redirect' => BASEURL . '/order/antrian/' . $result['nomor_antrian']
+                    ]);
+                    exit;
+                } else {
+                    throw new Exception('Checkout Gagal: ' . $result['message']);
+                }
+            } catch (Exception $e) {
+                // Clear output buffer
+                ob_end_clean();
+                
+                // Send JSON response
+                header('Content-Type: application/json');
+                echo json_encode([
+                    'status' => 'error',
+                    'message' => $e->getMessage()
+                ]);
+                exit;
             }
         }
     }
