+++ app/core/ErrorHandler.php
