+++ app/init.php
@@ -2,7 +2,18 @@
 if (session_status() === PHP_SESSION_NONE) {
     session_start();
 }
-require_once 'config/config.php';
-require_once 'core/App.php';
-require_once 'core/Controller.php';
-require_once 'core/Database.php';
+
+// Error handling setup
+require_once __DIR__ . '/core/ErrorHandler.php';
+$errorHandler = ErrorHandler::getInstance();
+set_error_handler([$errorHandler, 'handleError']);
+set_exception_handler([$errorHandler, 'handleException']);
+
+// Display errors in development environment
+ini_set('display_errors', 1);
+error_reporting(E_ALL);
+
+require_once __DIR__ . '/config/config.php';
+require_once __DIR__ . '/core/App.php';
+require_once __DIR__ . '/core/Controller.php';
+require_once __DIR__ . '/core/Database.php';
