--- kantinqueue/app/config/config.php
@@ -1,6 +1,0 @@
-<?php
-define('BASEURL', 'http://localhost/KantinQueue/public');
-define('DB_HOST', 'localhost');
-define('DB_USER', 'root');
-define('DB_PASS', '');     
-define('DB_NAME', 'kantin_queue');
