--- kantinqueue/app/views/home/index.php
@@ -1,7 +1,0 @@
-<div class="p-5 mb-4 bg-light rounded-3">
-  <div class="container-fluid py-5">
-    <h1 class="display-5 fw-bold">Selamat Datang, <?= $data['nama']; ?>!</h1>
-    <p class="col-md-8 fs-4">Selamat datang di KantinQueue. Pesan makanan favoritmu tanpa perlu antri panjang. Cek menu dan langsung pesan sekarang.</p>
-    <a href="<?= BASEURL; ?>/order" class="btn btn-primary btn-lg" type="button">Pesan Sekarang <i class="bi bi-arrow-right"></i></a>
-  </div>
-</div>
