--- kantinqueue/app/views/templates/footer.php
@@ -1,7 +1,0 @@
-</main>
-<footer class="text-center mt-5 py-3 bg-light">
-    <p class="mb-0">&copy; KantinQueue - Kelompok 9</p>
-</footer>
-<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"></script>
-</body>
-</html>
