--- kantinqueue/public/index.php
@@ -1,8 +1,0 @@
-<?php
-if (!session_id()) {
-    session_start();
-}
-
-require_once '../app/init.php';
-
-$app = new App;
